made apk, apt, brew updates conditional

+++ .zshrc
@@ -162,8 +162,24 @@ alias gen-ospf-key="dd if=/dev/urandom count=1024 | shasum"
 # Bulk rename of logs from RoyalTSX
 # For some reason, they don't let you change the file suffix
 alias rename_logs="autoload zmv;zmv -W '*.log' '*.txt'"
-alias bu="brew update && brew upgrade && brew upgrade --cask"
-alias au="sudo apt update && sudo apt full-upgrade -y && sudo apt autoremove -y"
+
+# Add homebrew update if availalble
+if _has brew;
+then
+    alias bu="brew update && brew upgrade && brew upgrade --cask"
+fi
+
+# Add apt update if availalble
+if _has apt;
+then
+    alias au="sudo apt update && sudo apt full-upgrade -y && sudo apt autoremove -y"
+fi
+
+# Add apk update if availalble
+if _has apk;
+then
+    alias au="sudo apk upgrade --update-cache"
+fi
 
 # Use colordiff if it's available
 if _has colordiff;
